requirements.txt
numpy~=1.24.2
opencv-python~=4.7.0.72
face_recognition
Pillow~=9.4.0
#apt-get update && apt-get install ffmpeg libsm6 libxext6  -y
